+++ rocketpy/motors/hybrid_motor.py
@@ -72,6 +72,8 @@ class HybridMotor(Motor):
     HybridMotor.propellant_mass : Function
         Total propellant mass in kg as a function of time, this includes the
         mass of fluids in each tank and the mass of the solid grains.
+    HybridMotor.structural_mass_ratio: float
+        Initial ratio between the dry mass and the total mass.
     HybridMotor.total_mass_flow_rate : Function
         Time derivative of propellant total mass in kg/s as a function
         of time as obtained by the thrust source.
@@ -599,16 +601,19 @@ class HybridMotor(Motor):
         )
         reset_funcified_methods(self)
 
-    def draw(self):
-        """Draws a representation of the HybridMotor."""
-        self.plots.draw()
+    def draw(self, *, filename=None):
+        """Draws a representation of the HybridMotor.
 
-    def info(self):
-        """Prints out basic data about the Motor."""
-        self.prints.all()
-        self.plots.thrust()
+        Parameters
+        ----------
+        filename : str | None, optional
+            The path the plot should be saved to. By default None, in which case
+            the plot will be shown instead of saved. Supported file endings are:
+            eps, jpg, jpeg, pdf, pgf, png, ps, raw, rgba, svg, svgz, tif, tiff
+            and webp (these are the formats supported by matplotlib).
 
-    def all_info(self):
-        """Prints out all data and graphs available about the Motor."""
-        self.prints.all()
-        self.plots.all()
+        Returns
+        -------
+        None
+        """
+        self.plots.draw(filename=filename)