+++ rocketpy/motors/liquid_motor.py
@@ -47,6 +47,8 @@ class LiquidMotor(Motor):
     LiquidMotor.propellant_mass : Function
         Total propellant mass in kg as a function of time, includes fuel
         and oxidizer.
+    LiquidMotor.structural_mass_ratio: float
+        Initial ratio between the dry mass and the total mass.
     LiquidMotor.total_mass_flow_rate : Function
         Time derivative of propellant total mass in kg/s as a function
         of time as obtained by the tanks mass flow.
@@ -461,21 +463,19 @@ class LiquidMotor(Motor):
         self.positioned_tanks.append({"tank": tank, "position": position})
         reset_funcified_methods(self)
 
-    def draw(self):
-        """Draw a representation of the LiquidMotor."""
-        self.plots.draw()
+    def draw(self, *, filename=None):
+        """Draw a representation of the LiquidMotor.
 
-    def info(self):
-        """Prints out basic data about the Motor."""
-        self.prints.all()
-        self.plots.thrust()
+        Parameters
+        ----------
+        filename : str | None, optional
+            The path the plot should be saved to. By default None, in which case
+            the plot will be shown instead of saved. Supported file endings are:
+            eps, jpg, jpeg, pdf, pgf, png, ps, raw, rgba, svg, svgz, tif, tiff
+            and webp (these are the formats supported by matplotlib).
 
-    def all_info(self):
-        """Prints out all data and graphs available about the Motor.
-
-        Return
-        ------
+        Returns
+        -------
         None
         """
-        self.prints.all()
-        self.plots.all()
+        self.plots.draw(filename=filename)